Move dialog variables to its own file
