+++ src/Rhisis.Core/Structures/Game/DialogVariables.cs
@@ -1,0 +1,10 @@
+﻿namespace Rhisis.Core.Structures.Game
+{
+    public class DialogVariables
+    {
+        /// <summary>
+        /// Player name variable for dialogs
+        /// </summary>
+        public const string PlayerNameText = "%PLAYERNAME%";
+    }
+}
